debug

+++ SaleManages/SaleManages/SaleManages/DAO/ObjBillDAO.cs
@@ -104,11 +104,17 @@ namespace SaleManages.DAO
         {
             System.Windows.Forms.Form f = System.Windows.Forms.Application.OpenForms["_frmSalesManage"];
             string SOHD = ((_frmSalesManage)f).tbCodeHD.Text;
-            string DeleteQuery = "DELETE FROM SANPHAM WHERE MASP = '" + SOHD + "'";
-            int result = DataProvider.Instance.ExecuteNonQuery(DeleteQuery);
-            if (result > 0)
+            string CheckQuery = "SELECT * " +
+                "FROM CTHD WHERE SOHD = '" + SOHD + "' ";
+            DataTable dt = DataProvider.Instance.ExecuteQuery(CheckQuery);
+            if (dt.Rows.Count == 0)
             {
-                MessageBox.Show("Hoá Đơn đã bị xoá,bấm Xem để xem dữ liệu mới", "Thông báo", MessageBoxButtons.OK);
+                string DeleteQuery = "DELETE FROM SANPHAM WHERE MASP = '" + SOHD + "'";
+                int result = DataProvider.Instance.ExecuteNonQuery(DeleteQuery);
+                if (result > 0)
+                {
+                    MessageBox.Show("Hoá Đơn đã bị xoá,bấm Xem để xem dữ liệu mới", "Thông báo", MessageBoxButtons.OK);
+                }
             }
         }
